# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SentinelLLM gateway contracts",
    "version": "0.3.0",
    "description": "Phase 3 gateway: buffered generation, content/tool/schema enforcement, durable tenant-scoped operation metadata, optimistic versioned policy/registry updates, and Redis-backed admission. Public gateway and private inspection/MCP services have distinct boundaries."
  },
  "paths": {
...
    "/internal/v1/inspect": {
      "post": {
        "summary": "Phase 1 private findings service",
        "security": [
          {
            "serviceBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/inspection-request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Approved result; preview may contain a denial decision",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inspection-response"
                }
              }
            },
            "headers": {
              "X-Sentinel-Operation-ID": {
                "description": "Server-generated operation UUID",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "400": {
            "description": "Sanitized error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            },
            "headers": {
              "X-Sentinel-Operation-ID": {
                "description": "Server-generated operation UUID",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "401": {
            "description": "Sanitized error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            },
            "headers": {
              "X-Sentinel-Operation-ID": {
                "description": "Server-generated operation UUID",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "403": {
            "description": "Sanitized error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            },
            "headers": {
              "X-Sentinel-Operation-ID": {
                "description": "Server-generated operation UUID",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "413": {
            "description": "Sanitized error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            },
            "headers": {
              "X-Sentinel-Operation-ID": {
                "description": "Server-generated operation UUID",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "429": {
            "description": "Sanitized error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            },
            "headers": {
              "X-Sentinel-Operation-ID": {
                "description": "Server-generated operation UUID",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "502": {
            "description": "Sanitized error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            },
            "headers": {
              "X-Sentinel-Operation-ID": {
                "description": "Server-generated operation UUID",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "503": {
            "description": "Sanitized error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            },
            "headers": {
              "X-Sentinel-Operation-ID": {
                "description": "Server-generated operation UUID",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "504": {
            "description": "Sanitized error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            },
            "headers": {
              "X-Sentinel-Operation-ID": {
                "description": "Server-generated operation UUID",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          }
        },
        "servers": [
          {
            "url": "http://inspection:8000",
            "description": "Private inspection service; not exposed by gateway"
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "applicationBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "serviceBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "inspection-request": {
        "$ref": "schemas/inspection-request.schema.json"
      },
      "inspection-response": {
        "$ref": "schemas/inspection-response.schema.json"
      },
...
      "error": {
        "$ref": "schemas/error.schema.json"
      },
...
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development gateway"
    }
  ]
}